代码重构
