+++ examples/defog/defog_search.py
@@ -21,19 +21,17 @@ import tensorlayerx as tlx
 
 sys.path.insert(0, os.path.dirname(os.path.abspath(__file__)))
 
+from gammagl.models.defog import DeFoGModel
+from extra_features import ExtraFeatures, DummyExtraFeatures
+from extra_features_molecular import ExtraMolecularFeatures, DummyMolecularFeatures
+from noise_distribution import NoiseDistribution
+from rate_matrix import RateMatrixDesigner
+from time_distorter import TimeDistorter
+from sampler import sample_batch
+from evaluator import evaluate_generated_graphs
 from defog_trainer import (
-    DeFoGModel,
-    DummyExtraFeatures,
-    DummyMolecularFeatures,
-    ExtraFeatures,
-    ExtraMolecularFeatures,
-    NoiseDistribution,
-    RateMatrixDesigner,
-    TimeDistorter,
-    evaluate_generated_graphs,
     load_model_snapshot_for_sampling,
     load_real_dataset,
-    sample_batch,
 )
