+++ examples/defog/defog_utils.py
@@ -395,46 +395,6 @@ class PlaceHolder:
 
 
 # ============================================================
-# Timestep Embedding
-# ============================================================
-
-def timestep_embedding(timesteps, dim, max_period=10000):
-    r"""Create sinusoidal timestep embeddings.
-
-    Parameters
-    ----------
-    timesteps : tensor
-        1D or 2D tensor of timestep values.
-    dim : int
-        Output embedding dimension.
-    max_period : int
-        Maximum period for the sinusoidal encoding.
-
-    Returns
-    -------
-    tensor
-        Embedding of shape ``(N, dim)``.
-    """
-    half = dim // 2
-    freqs = np.exp(-math.log(max_period) * np.arange(0, half, dtype=np.float32) / half)
-    freqs = tlx.convert_to_tensor(freqs)
-
-    ts = tlx.cast(tlx.reshape(timesteps, [-1, 1]), tlx.float32)
-    freqs = tlx.reshape(freqs, [1, -1])
-    args = ts * freqs  # (N, half)
-
-    cos_part = tlx.cos(args)
-    sin_part = tlx.sin(args)
-    embedding = tlx.concat([cos_part, sin_part], axis=-1)  # (N, dim) or (N, dim-1)
-
-    if dim % 2 == 1:
-        zeros_pad = tlx.zeros([embedding.shape[0], 1], dtype=tlx.float32)
-        embedding = tlx.concat([embedding, zeros_pad], axis=-1)
-
-    return embedding
-
-
-# ============================================================
 # Dense conversion utilities
 # ============================================================
